Changed to use helper methods
