+++ .gitignore
@@ -15,3 +15,4 @@ gh-pages
 # code coverage
 coverage
 
+*.orig
