+++ src/lib/utility/dom-helpers.js
@@ -23,3 +23,27 @@ export function getParentPosition(element) {
   }
   return { x: xPosition, y: yPosition }
 }
+
+export function getSumScroll(node) {
+  if (node === document.body) {
+    return {scrollLeft: 0, scrollTop: 0}
+  } else {
+    const parent = this.getSumScrolls(node.parentNode)
+    return ({
+      scrollLeft: node.scrollLeft + parent.scrollLeft,
+      scrollTop: node.scrollTop + parent.scrollTop
+    })
+  }
+}
+
+export function getSumOffset(node) {
+  if (node === document.body) {
+    return {offsetLeft: 0, offsetTop: 0}
+  } else {
+    const parent = this.getSumOffset(node.offsetParent)
+    return ({
+      offsetLeft: node.offsetLeft + parent.offsetLeft,
+      offsetTop: node.offsetTop + parent.offsetTop
+    })
+  }
+}
